requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Mail==0.10.0
gunicorn==23.0.0
requests==2.32.3
feedparser==6.0.11
Werkzeug==3.0.4
python-dotenv
livereload
